requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
jinja2>=3.1.0
python-multipart>=0.0.9
openai-whisper==20231117
google-genai>=0.2.0
openai>=1.10.0
anthropic>=0.18.0
feedparser>=6.0.10
pydantic-settings>=2.1.0
aiofiles>=23.2.0

piper-tts==1.2.0
